Remove this file
A request was received from @lgarron to use a different image to avoid any possible confusion (a completely fair request and valid point); should replace this at some point, but seeing as the point of the site is to provide an API, how it looks in a browser really doesn't matter, so just removing the file completely. Fixes #3.